Adds a board common API for retriving the SoC' Silicon revision data/errata
    This abstraction will support mcu's other than the stm32 family.
    It moves the systemlib/mcu_version.c functionality to
    common/stm32/board_mcu_version.c
